+++ packages/webfinger/tsdown.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from "tsdown";
 export default [
   defineConfig({
     entry: ["src/mod.ts"],
-    dts: true,
+    dts: { compilerOptions: { isolatedDeclarations: true, declaration: true } },
     format: ["esm", "cjs"],
     platform: "node",
     external: [/^node:/],